Add files via upload
All test files as well as most output for Microsoft Solver Foundation(Simplex Directive) as well as a binary branch-and-bound implementation with the best-first search heuristic in the Qlik Sense Analytics platform. The latter output files start with "QSA_", otherwise the name of the output file is the same as the corresponding input file. Some input files are missing output files since they took too long to compute, notably 500x500 and 1000x1000 examples as well as the hardest 200x200 example.